separate capabilities into files
