+++ features/access.feature
@@ -6,6 +6,7 @@ Feature: Access should not be given for private end-points unless identity is ve
     Scenario: Not being allowed access to a private end point when no id provided
 
         Given I am not identified
+        And I am not authorised
         When I POST the following to "/datasets/E3BC0B6-D6C4-4E20-917E-95D7EA8C91DC":
             """
             {
@@ -16,6 +17,7 @@ Feature: Access should not be given for private end-points unless identity is ve
 
     Scenario: Being allowed access to a private end point when id provided
         Given I am identified as "user@ons.gov.uk"
+        And I am authorised
         When I POST the following to "/datasets/E3BC0B6-D6C4-4E20-917E-95D7EA8C91DC":
             """
             {
