+++ features/steps/error_feature.go
@@ -1,0 +1,15 @@
+package steps_test
+
+import "fmt"
+
+type ErrorFeature struct {
+	err error
+}
+
+func (f *ErrorFeature) Errorf(format string, args ...interface{}) {
+	f.err = fmt.Errorf(format, args...)
+}
+
+func (f *ErrorFeature) StepError() error {
+	return f.err
+}
